interpreter: split keyword definitions out of the interpreter
